Create README.md
Not much explanation needed, this is an intuitive application. You select an answer and if your selection is correct, that selection will be highlighted in green.  If you are incorrect, it will highlight your selected answer in red and the correct answer in green. There is a pool of possible questions and the program will randomly choose a question that you have not yet been asked each time you press the next button.